급조한 주행 기록.

+++ app/src/main/java/kr/musekee/faremeter/components/main/ResultItem.kt
@@ -33,7 +33,7 @@ fun ResultItem(
     idx: Int,
     transportation: String,
     fare: Int,
-    distance: Float,
+    distance: Double,
     date: Date
 ) {
